backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
earthengine-api==0.1.365
google-auth==2.23.3
google-cloud-storage==2.10.0
requests==2.31.0
python-dotenv==1.0.0